eth2util/eth2exp: spec new proposal aggregate endpoint
